Sprite Col Fixes

+++ InputManager.hpp
@@ -13,7 +13,7 @@ namespace Alex
 		bool IsSpriteClicked(sf::Sprite object, sf::Mouse::Button button, sf::RenderWindow &window);
 		bool IsSpriteHovering(sf::Sprite object, sf::RenderWindow &window);
 
-		sf::Vector2i GetMousePosition(sf::RenderWindow &window);
+		sf::Vector2f GetMousePosition(sf::RenderWindow &window);
 
 	};
 }
